requirements.txt
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
dlib==19.18.0
idna==2.9
imagezmq==1.1.1
imutils==0.5.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pytesseract==0.3.2
pyzmq==18.1.1
requests==2.23.0
urllib3==1.25.8
zmq==0.0.0
